Tests/LibWeb/Text/expected/css/insert-rule-constraints.txt

Inserting import rules after style rule throws: HierarchyRequestError
Inserting namespace rules after style rule throws: HierarchyRequestError
Inserting namespace rules before import rule throws: HierarchyRequestError
Inserting style rules before import rule throws: HierarchyRequestError
Inserting style rules before namespace rule throws: HierarchyRequestError
